Renaming due to replacing JWT/jose with iron-session
