+++ .env.example
@@ -12,7 +12,7 @@ SAML_ENTITY_ID=https://churro-test.stanford.edu
 # JWT Configuration
 # Generate a secure secret with: openssl rand -base64 32
 # REQUIRED - Application will not start without this
-JWT_SECRET=
+SESSION_SECRET=
 
 # Acquia Cloud API Configuration
 # Get these from Acquia Cloud Platform
